fix system_stats type

+++ include/system.h
@@ -4,12 +4,12 @@
 #define CONVERT_KB_TO_GB 1048576
 
 struct cpu_stats {
-    unsigned long user, nice, system, idle, iowait, irq, softirq, steal, guest, guest_nice;
-    unsigned long total, idle_time;
+    unsigned long long user, nice, system, idle, iowait, irq, softirq, steal, guest, guest_nice;
+    unsigned long long total, idle_time;
 };
 
 struct mem_stats {
-    unsigned long mem_total, mem_available;
+    unsigned long long mem_total, mem_available;
 };
 
 struct system_stats {
